software/experiment_scheduler/requirements.txt
coloredlogs==15.0.1
numpy~=1.26.4
netifaces==0.11.0
typing_extensions~=4.12.2
serial
pyserial~=3.5

pandas~=2.2.3
matplotlib~=3.10.1
tqdm~=4.67.1
psutil~=7.0.0
PyYAML~=6.0.2
requests~=2.32.3
jupyter~=1.1.1
pip~=23.2.1
attrs~=25.3.0
setuptools~=65.5.0
pytz~=2025.2
overrides~=7.7.0
cffi~=1.17.1
humanfriendly~=10.0
nats-py
